+++ spicerack/tests/unit/test_init.py
@@ -14,10 +14,12 @@ from requests.auth import HTTPBasicAuth
 from wmflib.actions import ActionsDict
 from wmflib.config import load_yaml_config
 from wmflib.dns import Dns
+from wmflib.interactive import AbortError
 from wmflib.phabricator import Phabricator
 from wmflib.prometheus import Prometheus, Thanos
 
 from spicerack import Spicerack
+from spicerack._cookbook import get_cookbook_callback
 from spicerack.administrative import Reason
 from spicerack.alerting import AlertingHosts
 from spicerack.alertmanager import Alertmanager, AlertmanagerHosts
@@ -29,7 +31,7 @@ from spicerack.debmonitor import Debmonitor
 from spicerack.dhcp import DHCP
 from spicerack.dnsdisc import Discovery
 from spicerack.elasticsearch_cluster import ElasticsearchClusters
-from spicerack.exceptions import SpicerackError
+from spicerack.exceptions import RunCookbookError, SpicerackError
 from spicerack.ganeti import Ganeti
 from spicerack.icinga import IcingaHosts
 from spicerack.ipmi import Ipmi
@@ -107,9 +109,15 @@ def test_spicerack(mocked_dns_resolver, monkeypatch):
         EtcdctlController,
     )
     assert isinstance(spicerack.kafka(), Kafka)
+
     service_catalog = spicerack.service_catalog()
     assert isinstance(service_catalog, Catalog)
     assert spicerack.service_catalog() is service_catalog  # Returned the cached instance
+    service_catalog_new = spicerack.service_catalog(refresh=True)
+    assert isinstance(service_catalog_new, Catalog)
+    assert service_catalog_new is not service_catalog
+    assert spicerack.service_catalog() is service_catalog_new  # Returned the new cached instance
+
     assert isinstance(spicerack.apt_get(mock.MagicMock(spec_set=RemoteHosts)), AptGetHosts)
     assert isinstance(spicerack.orchestrator(), Orchestrator)
     assert isinstance(spicerack.lock(), NoLock)
@@ -292,6 +300,28 @@ def test_run_cookbook_no_callback():
     spicerack = Spicerack(verbose=True, dry_run=False, **SPICERACK_TEST_PARAMS)
     with pytest.raises(SpicerackError, match="Unable to run other cookbooks, get_cookbook_callback is not set."):
         spicerack.run_cookbook("class_api.example", [])
+
+
+def test_run_cookbook_raise():
+    """It should raise a RunCookbookError if raises is set to True and the cookbook raises an exception."""
+    get_cookbook = get_cookbook_callback([get_fixture_path("cookbook")])
+    spicerack = Spicerack(verbose=True, dry_run=False, get_cookbook_callback=get_cookbook, **SPICERACK_TEST_PARAMS)
+    with pytest.raises(
+        RunCookbookError,
+        match="run_cookbook returned exit code 99 when running cookbook group3.raise_exception with args",
+    ):
+        spicerack.run_cookbook("group3.raise_exception", [], raises=True)
+
+
+@mock.patch("wmflib.interactive.ask_input", return_value="abort")
+def test_run_cookbook_confirm(mocked_ask_input):
+    """It should ask the user confirmation if confirm is set to True and the cookbook raises an exception."""
+    get_cookbook = get_cookbook_callback([get_fixture_path("cookbook")])
+    spicerack = Spicerack(verbose=True, dry_run=False, get_cookbook_callback=get_cookbook, **SPICERACK_TEST_PARAMS)
+    with pytest.raises(AbortError, match="Task manually aborted"):
+        spicerack.run_cookbook("group3.raise_exception", [], confirm=True)
+
+    mocked_ask_input.assert_called_once()
 
 
 @mock.patch("spicerack.Path.is_dir")